+++ plugins/clients/apps-client/src/hopeit/apps_client/__init__.py
@@ -14,7 +14,7 @@ import random
 import aiohttp
 from stringcase import spinalcase  # type: ignore
 
-from hopeit.app.client import AppConnectionNotFound, Client, ClientException
+from hopeit.app.client import AppConnectionNotFound, Client, ClientException, UnhandledResponse
 from hopeit.app.context import EventContext
 from hopeit.app.config import AppConfig, AppDescriptor, EventConnection, EventConnectionType
 from hopeit.app.errors import Unauthorized
@@ -237,17 +237,20 @@ class AppsClient(Client):
 
     async def call(self, event_name: str,
                    *, datatype: Type[EventPayloadType], payload: Optional[EventPayload],
-                   context: EventContext, **kwargs) -> List[EventPayloadType]:
+                   context: EventContext, responses: Optional[Dict[int, Type[EventPayloadType]]] = None,
+                   **kwargs) -> List[EventPayloadType]:
         """
         Invokes event on external app linked in config `app_connections` section.
         Target event must also be configured in event `client` section
 
-        :param: app_connection, str: name of configured app_connection to connect to
-        :param: event_name, str: target event name to inoke, configured in events section
-        :datatype: Type[EventPayloadType]: expected return type
-        :payload: optional payload to send when calling external event
-        :context: EventContext of current application
-        **kwargs: any other argument to be sent as query args when calling event
+        :param app_connection, str: name of configured app_connection to connect to
+        :param event_name, str: target event name to inoke, configured in events section
+        :param datatype: Type[EventPayloadType]: expected return type for 200 status response
+        :param payload: optional payload to send when calling external event
+        :param context: EventContext of current application
+        :param responses: Optional[Dict[int, Type[EventPayloadType]]] to handle non 200 status responses
+        ex.: {404: NotFoundResultClass}
+        :param **kwargs: any other argument to be sent as query args when calling event
 
         :return: datatype, returned data from invoked event, converted to datatype
         """
@@ -276,7 +279,7 @@ class AppsClient(Client):
                 if event_info.type == EventConnectionType.GET:
                     request_func = self.session.get(url, headers=headers, params=kwargs)
                     return await self._request(
-                        request_func, context, datatype, event_name, host_index
+                        request_func, context, datatype, event_name, host_index, responses
                     )
 
                 if event_info.type == EventConnectionType.POST:
@@ -284,7 +287,7 @@ class AppsClient(Client):
                         url, headers=headers, data=Payload.to_json(payload), params=kwargs
                     )
                     return await self._request(
-                        request_func, context, datatype, event_name, host_index
+                        request_func, context, datatype, event_name, host_index, responses
                     )
 
                 raise NotImplementedError(f"Event type {event_info.type.value} not supported")
@@ -394,29 +397,37 @@ class AppsClient(Client):
 
     async def _parse_response(self, response, context: EventContext,
                               datatype: Type[EventPayloadType],
-                              target_event_name: str) -> List[EventPayloadType]:
+                              target_event_name: str,
+                              responses: Optional[Dict[int, Type[EventPayloadType]]]) -> List[EventPayloadType]:
         """
         Parses http response from external App, catching Unathorized errors
         and converting the result to the desired datatype
         """
-        if response.status == 200:
+        try:
+            response_type = datatype if response.status == 200 else responses[response.status]  # type: ignore
+        except (TypeError, KeyError):
+            if response.status == 401:
+                raise Unauthorized(context.app_key)  # pylint: disable=raise-missing-from
+            if response.status >= 500:
+                raise ServerException(await response.text())  # pylint: disable=raise-missing-from
+            raise UnhandledResponse(  # pylint: disable=raise-missing-from
+                f"Missing {response.status} status handler, use `responses` to handle this exception",
+                await response.text(), response.status)
+
+        if response.content_type == 'text/plain':
+            data = {target_event_name: await response.text()}
+        else:
             data = await response.json()
-            if isinstance(data, list):
-                return Payload.from_obj(data, list, item_datatype=datatype)  # type: ignore
-            return [Payload.from_obj(data, datatype, key=target_event_name)]
-        if response.status == 401:
-            raise Unauthorized(context.app_key)
-        if response.status >= 500:
-            raise ServerException(await response.text())
-        raise RuntimeError(await response.text())
+        if isinstance(data, list):
+            return Payload.from_obj(data, list, item_datatype=response_type)  # type: ignore
+        return [Payload.from_obj(data, response_type, key=target_event_name)]
 
     async def _request(self, request_func: AbstractAsyncContextManager, context: EventContext,
-                       datatype: Type[EventPayloadType],
-                       target_event_name: str,
-                       host_index: int) -> List[EventPayloadType]:
+                       datatype: Type[EventPayloadType], target_event_name: str, host_index: int,
+                       responses: Optional[Dict[int, Type[EventPayloadType]]]) -> List[EventPayloadType]:
         async with request_func as response:
             result = await self._parse_response(
-                response, context, datatype, target_event_name
+                response, context, datatype, target_event_name, responses
             )
             self.conn_state.load_balancer.success(host_index)  # type: ignore
             return result