+++ docs/source/release-notes.rst
@@ -1,11 +1,21 @@
 Release Notes
 =============
 
+Version 0.16.4
+______________
+Plugin: 
+  - apps-client: `app_call` and `app_call_list` add the optional `responses` parameter, this allows to handle
+  different datatypes per response status code. Also the `UnhandledResponse` exception is added, when an unexpected
+  response is received it raises an `UnhandledResponse` enriched with information such as the `status` code and
+  `response` content of the request.
+  
+
 Version 0.16.3
 ______________
 Engine:
 - Fix: Calling the `hopeit_server` command line without `--enabled-groups` parameter or with an empty one prevents to
 start the server. Now `--enabled-groups` is an optional parameter.
+
 
 Version 0.16.2
 ______________